refactor(MobileSamoleTest.java): moved dragAndDrop method to independent class and suite
