+++ src/test/resources/testng_suites/android-dragAndDrop.xml
@@ -1,0 +1,26 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite name="Carina Demo Tests - Mobile Drag And Drop Sample (Android)" verbose="1" parallel="methods">
+
+    <parameter name="suiteOwner" value="qpsdemo" />
+    <parameter name="provider" value="mcloud"/>
+    <parameter name="jenkinsJobName" value="Mobile-Android-DragAndDrop-Test" />
+    <parameter name="jenkinsJobType" value="android"/>
+    <parameter name="jenkinsEmail" value="demo@qaprosoft.com" />
+    <parameter name="jenkinsEnvironments" value="DEMO" />
+
+    <parameter name="stringParam::capabilities.app::path drag and drop mobile application" value="https://qaprosoft.s3-us-west-2.amazonaws.com/carinademoexample.apk" />
+
+    <parameter name="jenkinsMobileDefaultPool" value="ANY" />
+
+    <parameter name="jenkinsRegressionPipeline" value="mobile_regression" />
+    <parameter name="jenkinsJobExecutionOrder" value="3" />
+    <parameter name="jenkinsJobExecutionMode" value="continue" />
+    <parameter name="jenkinsAutoScreenshot" value="true" />
+
+    <test name="Mobile drag and drop test">
+        <classes>
+            <class name="com.qaprosoft.carina.demo.MobileDragAndDropTest" />
+        </classes>
+    </test>
+
+</suite>
